save chatbot changes
